+++ lib/app/modules/report_list/bindings/report_list_binding.dart
@@ -1,0 +1,11 @@
+import 'package:get/get.dart';
+import '../controllers/report_list_controller.dart';
+
+class ReportListBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<ReportListController>(
+      () => ReportListController(),
+    );
+  }
+}