+++ lib/app/routes/app_pages.dart
@@ -5,6 +5,8 @@ import '../modules/home/bindings/home_binding.dart';
 import '../modules/home/views/home_view.dart';
 import '../modules/dynamic_form/bindings/dynamic_form_binding.dart';
 import '../modules/dynamic_form/views/dynamic_form_view.dart';
+import '../modules/report_list/bindings/report_list_binding.dart';
+import '../modules/report_list/views/report_list_view.dart';
 import '../modules/report_history/bindings/report_history_binding.dart';
 import '../modules/report_history/views/report_history_view.dart';
 import '../modules/report_detail/bindings/report_detail_binding.dart';
@@ -29,6 +31,11 @@ class AppPages {
       name: Routes.DYNAMIC_FORM,
       page: () => const DynamicFormView(),
       binding: DynamicFormBinding(),
+    ),
+    GetPage(
+      name: Routes.REPORT_LIST,
+      page: () => const ReportListView(),
+      binding: ReportListBinding(),
     ),
     GetPage(
       name: Routes.REPORT_HISTORY,